Move Api to right folder
