+++ src/Homes/index.js
@@ -6,7 +6,7 @@ import Filters from './Filters';
 import Cards from './Cards';
 import Header from '../Header';
 import GoogleMap from './GoogleMap';
-import getHomes from './Cards/api';
+import getHomes from './api';
 
 const Content = styled.main``;
